requirements.txt
discord.py>=2.1
aiomysql>=0.1.1
Pillow>=9.2.0
